requirements.txt
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0
cryptography==42.0.1
structlog==24.1.0 